Make withdraw lifecycle atomic

+++ backend/app/schemas/withdraw.py
@@ -1,9 +1,11 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class WithdrawCreate(BaseModel):
     telegram_id: int
-    amount: float
+    amount: Decimal
     card_number: str
     card_holder: str
     bank_name: str
